Split LogStore
